refactor(backend): extract admin user boundary
